Refactoring and cleanup for easier debugging and better code quality.
