+++ amr.py
@@ -1,5 +1,5 @@
-# -*- coding: utf-8 -*-
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
 
 """
 AMR (Abstract Meaning Representation) structure
@@ -145,10 +145,10 @@ class AMR(object):
             lines.append("Node "+ str(i) + " " + self.nodes[i])
             lines.append("Value: " + self.node_values[i])
             lines.append("Relations:")
-            for k, v in self.relations[i].items():
-                lines.append("Node " + v + " via " + k)
-            for k2, v2 in self.attributes[i].items():
-                lines.append("Attribute: " + k2 + " value " + v2)
+            for relation in self.relations[i]:
+                lines.append("Node " + relation[1] + " via " + relation[0])
+            for attribute in self.attributes[i]:
+                lines.append("Attribute: " + attribute[0] + " value " + attribute[1])
         return "\n".join(lines)
 
     def __repr__(self):
@@ -161,6 +161,32 @@ class AMR(object):
         """
         print >> DEBUG_LOG, self.__str__()
 
+    @staticmethod
+    def get_amr_line(input_f):
+        """
+        Read the file containing AMRs. AMRs are separated by a blank line.
+        Each call of get_amr_line() returns the next available AMR (in one-line form).
+        Note: this function does not verify if the AMR is valid
+
+        """
+        cur_amr = []
+        has_content = False
+        for line in input_f:
+            line = line.strip()
+            if line == "":
+                if not has_content:
+                    # empty lines before current AMR
+                    continue
+                else:
+                    # end of current AMR
+                    break
+            if line.strip().startswith("#"):
+                # ignore the comment line (starting with "#") in the AMR file
+                continue
+            else:
+                has_content = True
+                cur_amr.append(line.strip())
+        return "".join(cur_amr)
 
     @staticmethod
     def parse_AMR_line(line):
@@ -396,17 +422,19 @@ class AMR(object):
         return result_amr
 
 # test AMR parsing
+# run by
 # a unittest can also be used.
 if __name__ == "__main__":
     if len(sys.argv) < 2:
         print >> ERROR_LOG, "No file given"
         exit(1)
     amr_count = 1
-    for line in open(sys.argv[1]):
-        cur_line = line.strip()
-        if cur_line == "" or cur_line.startswith("#"):
-            continue
-        print >> DEBUG_LOG, "AMR", amr_count
-        current = AMR.parse_AMR_line(cur_line)
-        current.output_amr()
-        amr_count += 1
+    with open(sys.argv[1]) as input_f:
+        while True:
+            cur_line = AMR.get_amr_line(input_f)
+            if cur_line == "":
+                break
+            print >> DEBUG_LOG, "AMR", amr_count
+            current = AMR.parse_AMR_line(cur_line)
+            current.output_amr()
+            amr_count += 1
